[refactor] : 패키지, 디렉토리 네이밍 방식 변경 (#8)
