+++ lib/data/app_exceptions.dart
