+++ lib/view/example/components/app_bar.dart
